requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
coloredlogs==10.0
decorator==4.4.0
humanfriendly==4.18
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==7.0.0
numpy==1.16.2
pandas==0.24.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pytest==4.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1.1
six==1.12.0
stevedore==1.30.1
traitlets==4.3.2
virtualenv==16.4.3
virtualenv-clone==0.5.2
virtualenvwrapper==4.8.4
wcwidth==0.1.7
xlrd==1.2.0
